requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
aiofiles>=23.0.0

